% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymNmix-package.R
\docType{package}
\name{asymNmix-package}
\alias{asymNmix}
\alias{asymNmix-package}
\title{asymNmix: Asymptotic N-mixture Model Fitting}
\description{
asymNmix is an R package made to aid in fitting asymptotic N-mixture models.
}
\author{
\strong{Maintainer}: Matthew RP Parker \email{mrparker909@gmail.com} (\href{https://orcid.org/0000-0003-3021-7959}{ORCID})

}
\keyword{internal}
