% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{Ax_log}
\alias{Ax_log}
\title{FUNCTION_TITLE}
\usage{
Ax_log(logA, logx)
}
\arguments{
\item{logA}{PARAM_DESCRIPTION}

\item{logx}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
