% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{log_tp_MAT_lse}
\alias{log_tp_MAT_lse}
\title{FUNCTION_TITLE}
\usage{
log_tp_MAT_lse(M, omeg, gamm, corrections)
}
\arguments{
\item{M}{PARAM_DESCRIPTION}

\item{omeg}{PARAM_DESCRIPTION}

\item{gamm}{PARAM_DESCRIPTION}

\item{corrections}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[foreach]{foreach}}
}
