% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{logSumExp}
\alias{logSumExp}
\title{FUNCTION_TITLE}
\usage{
logSumExp(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
