% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{Pab}
\alias{Pab}
\title{FUNCTION_TITLE}
\usage{
Pab(a, b, omega, gamma)
}
\arguments{
\item{a}{PARAM_DESCRIPTION}

\item{b}{PARAM_DESCRIPTION}

\item{omega}{PARAM_DESCRIPTION}

\item{gamma}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
