% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{Pab_asymptotic}
\alias{Pab_asymptotic}
\title{FUNCTION_TITLE}
\usage{
Pab_asymptotic(a, b, omega, gamma, corrections = FALSE)
}
\arguments{
\item{a}{PARAM_DESCRIPTION}

\item{b}{PARAM_DESCRIPTION}

\item{omega}{PARAM_DESCRIPTION}

\item{gamma}{PARAM_DESCRIPTION}

\item{corrections}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
