% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymNmix-package.R
\docType{package}
\name{quickNmix-package}
\alias{quickNmix}
\alias{quickNmix-package}
\title{quickNmix: Asymptotic N-mixture Model Fitting}
\description{
quickNmix is an R package made to aid in fitting asymptotic N-mixture models.
}
\author{
\strong{Maintainer}: Matthew RP Parker \email{mrparker909@gmail.com} (\href{https://orcid.org/0000-0003-3021-7959}{ORCID})

Authors:
\itemize{
  \item Lloyd Elliott
  \item Laura LE Cowen
  \item Jiguo Cao
}

}
\keyword{internal}
