% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymNmix-package.R
\docType{package}
\name{quickNmix-package}
\alias{quickNmix}
\alias{quickNmix-package}
\title{quickNmix: Asymptotic N-mixture Model Fitting}
\description{
quickNmix aids in fitting asymptotic N-mixture models.
}
\author{
\strong{Maintainer}: Matthew RP Parker \email{mrparker909@gmail.com} (\href{https://orcid.org/0000-0003-3021-7959}{ORCID})

Authors:
\itemize{
  \item Lloyd Elliott \email{lloyd_elliott@sfu.ca} (\href{https://orcid.org/0000-0003-2187-7314}{ORCID})
  \item Laura LE Cowen \email{lcowen@uvic.ca} (\href{https://orcid.org/0000-0002-0853-1450}{ORCID})
  \item Jiguo Cao \email{jiguo_cao@sfu.ca} (\href{https://orcid.org/0000-0001-7417-6330}{ORCID})
}

}
\keyword{internal}
